Frontend/requirements.txt
requests==2.32.3
streamlit==1.44.1
python-dotenv==1.0.1